requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
faiss-cpu>=1.8.0
tiktoken>=0.7.0
streamlit>=1.36.0
matplotlib~=3.7.1